Json serialization merge resolution (#74)
* Deletes redundant frmScriptBuilder.

* Replaces references to .xml when checking file names with .json.

* Replaces references to .xml with .json.

* taskt.Core.Common.Clone now uses NewtonSoftJson
After the swap from XML serialization to Json serialization several features stopped working
due to Clone() relying on XML includes and other such properties. It is now rewritten to have a
dynamic return type and use NewtonSoftJson to create the clone instead of a MemoryStreamWriter.

* Updates the usages of taskt.Core.Common.Clone
When Clone() was updated some of the typing was made unnecessary because of the new dynamic
return type. The usages were revised and editted.

* Removes .xml reference in MessageBox